Starting LDAP coverage test
